Add documentation, enhance CMake presets, and refactor tensor handling
- Introduced a development guide for the Portable Tensor project.
- Updated CMake presets to include manifest features and improved Windows setup.
- Refactored tensor-related code to use new wrapper functions for better type handling.
- Removed deprecated tensor casting functions and improved dtype handling.
- Enhanced error handling in tensor comparison tests.

+++ native/cpp/core/src/tensor.cpp
@@ -111,7 +111,7 @@ PtensorResult<Tensor> Tensor::to_contiguous() const {
             for (size_t i = 0; i < num_elements; i++) {
                 int64_t from_index = 0;
 
-                for (int j = 0; j < ndims; j++) {
+                for (size_t j = 0; j < ndims; j++) {
                     from_index += coords[j] * this_stride[j];
                 }
